Lumberyard Release 1.24.0.0
        Binaries Uncompressed Size: 22.49GB (24146043221 bytes)

+++ dev/Code/Sandbox/Plugins/CryDesigner/Tools/Shape/CurveTool.cpp
@@ -207,7 +207,7 @@ void CurveTool::PrepareBeizerSpots(CViewport* view, UINT nFlags, const QPoint& p
 
 void CurveTool::Display(DisplayContext& dc)
 {
-    int oldThickness = dc.GetLineWidth();
+    float oldThickness = dc.GetLineWidth();
 
     if (m_ArcState == eArcState_ChooseFirstPoint || m_ArcState == eArcState_ChooseLastPoint)
     {
